requirements.txt
Flask==0.10.1
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.19
Pillow==2.3.0
SQLAlchemy==0.9.3
Werkzeug==0.9.4
alembic==0.6.3
bcrypt==1.0.2
cffi==0.8.2
itsdangerous==0.23
pycparser==2.10
wsgiref==0.1.2
